+++ packages/angular/components/details/details.component.ts
@@ -46,6 +46,8 @@ import '@synergy-design-system/components/components/details/details.js';
  *
  * @animation details.show - The animation to use when showing details. You can use `height: auto` with this animation.
  * @animation details.hide - The animation to use when hiding details. You can use `height: auto` with this animation.
+ *
+ * @cssproperty --syn-details-open-rotation - The rotation angle of the summary icon when the details is open.
  */
 @Component({
   selector: 'syn-details',