+++ packages/components/src/components/divider/divider.custom.styles.ts
@@ -2,7 +2,4 @@ import { css } from 'lit';
 
 export default css`
   /* Write custom CSS here */
-  :host {
-    --color: var(--syn-color-neutral-400);
-  }
 `;