+++ packages/components/src/components/input/input.component.ts
@@ -196,7 +196,15 @@ export default class SynInput extends SynergyElement implements SynergyFormContr
   @property() autocapitalize: 'off' | 'none' | 'on' | 'sentences' | 'words' | 'characters';
 
   /** Indicates whether the browser's autocorrect feature is on or off. */
-  @property() autocorrect: 'off' | 'on';
+  @property({
+    attribute: 'autocorrect',
+    reflect: true,
+    converter: {
+      fromAttribute: (value: string) => value === '' || value === 'on',
+      toAttribute: (value: boolean) => (value ? 'on' : 'off')
+    },
+    type: Boolean,
+  }) autocorrect: boolean;
 
   /**
    * Specifies what permission the browser has to provide assistance in filling out form field values. Refer to
@@ -832,7 +840,7 @@ export default class SynInput extends SynergyElement implements SynergyFormContr
               .value=${live(this.value)}
               autocapitalize=${ifDefined(this.autocapitalize)}
               autocomplete=${ifDefined(this.autocomplete)}
-              autocorrect=${ifDefined(this.autocorrect)}
+              autocorrect=${ifDefined(this.autocorrect ? undefined : 'off')}
               ?autofocus=${this.autofocus}
               spellcheck=${this.spellcheck}
               pattern=${ifDefined(this.pattern)}